consolidate python plugins
